Tests passing and start notebook

+++ examples/events_to_instantaneous_frames.rs
@@ -19,9 +19,12 @@ fn main() {
     let output_path = "/home/andrew/Downloads/temppp_out";
     let mut output_stream = BufWriter::new(File::create(output_path).unwrap());
 
-    let reconstructed_frame_rate = 500;
+    let reconstructed_frame_rate = 500.0;
     // For instantaneous reconstruction, make sure the frame rate matches the source video rate
-    assert_eq!(stream.tps / stream.ref_interval, reconstructed_frame_rate);
+    assert_eq!(
+        stream.tps / stream.ref_interval,
+        reconstructed_frame_rate as u32
+    );
 
     let mut frame_sequence: FrameSequence<u8> = FramerBuilder::new(
         stream.height.into(),
